Generate a proper service certificate for the autoscaler service.
Fixes SRVKS-186

Today, the autoscaler's custom-metrics API will start to fail running a few hours on an Openshift cluster. That's likely due to some form of certificate rotation breaking it.

Based on the documentation for the prometheus adapter (which is another implementation of the custom-metrics API) here: https://docs.openshift.com/container-platform/4.1/monitoring/exposing-custom-application-metrics-for-autoscaling.html this generates proper serving certificates for the API server to use and hooks them up in the existing manifests.