Enable platforms to filter resources and drop the custom-metrics APIService.
Fixes SRVKS-210.

This enables each platform/plugin to define a number of filters which are applied to the list of read resources before we apply any transformers to them, let alone installing them. Effectively, this enables us to drop certain resources programmatically.

The APIService for the custom-metrics API is dropped for now to not trigger SRVKS-186 unnecessarily.