Add network policies.
Fixes SRVKS-206

Newer ServiceMesh versions will tighten up the default network policy so we need to explictly tell it to allow traffic to flow into our controllers. In case of webhook, activator and autoscaler, traffic needs to be able to flow in for these components to actually work. For the remaining components, we want to be able to scrape metrics from them, so traffic needs to be allowed in as well.

In the future we could look at tightening those policies down even more, i.e. to only allow prometheus traffic to hit the controllers etc.